+++ template-skeleton/template/src/components/common/LazyLoadRoute.js
@@ -1,6 +1,6 @@
 import { Suspense } from 'react';
 import TopLoading from './TopLoading';
 
-export default function LazyLoadRoute({ children }) {
-  return <Suspense fallback={<TopLoading />}>{children}</Suspense>;
+export default function LazyLoadRoute({ children, fallback = <TopLoading /> }) {
+  return <Suspense fallback={fallback}>{children}</Suspense>;
 }